+++ internal/provider/provider.go
@@ -5,34 +5,43 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
+
+	"github.com/terraform-providers/terraform-provider-random/internal/resources/id"
+	"github.com/terraform-providers/terraform-provider-random/internal/resources/integer"
+	"github.com/terraform-providers/terraform-provider-random/internal/resources/password"
+	"github.com/terraform-providers/terraform-provider-random/internal/resources/pet"
+	"github.com/terraform-providers/terraform-provider-random/internal/resources/shuffle"
+	stringresource "github.com/terraform-providers/terraform-provider-random/internal/resources/string"
+	"github.com/terraform-providers/terraform-provider-random/internal/resources/uuid"
 )
 
-type provider struct {
-}
-
-func NewFramework() tfsdk.Provider {
+func NewProvider() tfsdk.Provider {
 	return &provider{}
 }
 
-func (p *provider) GetSchema(ctx context.Context) (tfsdk.Schema, diag.Diagnostics) {
+var _ tfsdk.Provider = (*provider)(nil)
+
+type provider struct{}
+
+func (p *provider) GetSchema(context.Context) (tfsdk.Schema, diag.Diagnostics) {
 	return tfsdk.Schema{}, nil
 }
 
-func (p *provider) Configure(ctx context.Context, req tfsdk.ConfigureProviderRequest, resp *tfsdk.ConfigureProviderResponse) {
+func (p *provider) Configure(context.Context, tfsdk.ConfigureProviderRequest, *tfsdk.ConfigureProviderResponse) {
 }
 
-func (p *provider) GetResources(ctx context.Context) (map[string]tfsdk.ResourceType, diag.Diagnostics) {
+func (p *provider) GetResources(context.Context) (map[string]tfsdk.ResourceType, diag.Diagnostics) {
 	return map[string]tfsdk.ResourceType{
-		"random_id":       resourceIDType{},
-		"random_integer":  resourceIntegerType{},
-		"random_password": resourcePasswordType{},
-		"random_pet":      resourcePetType{},
-		"random_shuffle":  resourceShuffleType{},
-		"random_string":   resourceStringType{},
-		"random_uuid":     resourceUUIDType{},
+		"random_id":       id.NewResourceType(),
+		"random_integer":  integer.NewResourceType(),
+		"random_password": password.NewResourceType(),
+		"random_pet":      pet.NewResourceType(),
+		"random_shuffle":  shuffle.NewResourceType(),
+		"random_string":   stringresource.NewResourceType(),
+		"random_uuid":     uuid.NewResourceType(),
 	}, nil
 }
 
-func (p *provider) GetDataSources(ctx context.Context) (map[string]tfsdk.DataSourceType, diag.Diagnostics) {
+func (p *provider) GetDataSources(context.Context) (map[string]tfsdk.DataSourceType, diag.Diagnostics) {
 	return map[string]tfsdk.DataSourceType{}, nil
 }
