+++ internal/provider/provider_resource_id_test.go
@@ -8,7 +8,6 @@ import (
 
 func TestAccResourceID(t *testing.T) {
 	resource.UnitTest(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories(),
 		Steps: []resource.TestStep{
 			{
@@ -33,7 +32,6 @@ func TestAccResourceID(t *testing.T) {
 
 func TestAccResourceID_importWithPrefix(t *testing.T) {
 	resource.UnitTest(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories(),
 		Steps: []resource.TestStep{
 			{
