+++ internal/provider/provider_test.go
@@ -1,20 +1,15 @@
 package provider
 
 import (
-	"testing"
-
 	"github.com/hashicorp/terraform-plugin-framework/providerserver"
 	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
 )
-
-func testAccPreCheck(t *testing.T) {
-}
 
 //nolint:unparam
 func testAccProtoV6ProviderFactories() map[string]func() (tfprotov6.ProviderServer, error) {
 	return map[string]func() (tfprotov6.ProviderServer, error){
 		"random": func() (tfprotov6.ProviderServer, error) {
-			return providerserver.NewProtocol6(NewFramework())(), nil
+			return providerserver.NewProtocol6(NewProvider())(), nil
 		},
 	}
 }
