+++ internal/resources/shuffle/resource.go
@@ -1,4 +1,4 @@
-package provider
+package shuffle
 
 import (
 	"context"
@@ -7,11 +7,19 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/hashicorp/terraform-plugin-framework/types"
+
+	"github.com/terraform-providers/terraform-provider-random/internal/random"
 )
 
-type resourceShuffleType struct{}
+func NewResourceType() *resourceType {
+	return &resourceType{}
+}
 
-func (r resourceShuffleType) GetSchema(context.Context) (tfsdk.Schema, diag.Diagnostics) {
+var _ tfsdk.ResourceType = (*resourceType)(nil)
+
+type resourceType struct{}
+
+func (r *resourceType) GetSchema(context.Context) (tfsdk.Schema, diag.Diagnostics) {
 	return tfsdk.Schema{
 		Description: "The resource `random_shuffle` generates a random permutation of a list of strings " +
 			"given as an argument.",
@@ -22,8 +30,10 @@ func (r resourceShuffleType) GetSchema(context.Context) (tfsdk.Schema, diag.Diag
 				Type: types.MapType{
 					ElemType: types.StringType,
 				},
-				Optional:      true,
-				PlanModifiers: []tfsdk.AttributePlanModifier{tfsdk.RequiresReplace()},
+				Optional: true,
+				PlanModifiers: []tfsdk.AttributePlanModifier{
+					tfsdk.RequiresReplace(),
+				},
 			},
 			"seed": {
 				Description: "Arbitrary string with which to seed the random number generator, in order to " +
@@ -32,26 +42,32 @@ func (r resourceShuffleType) GetSchema(context.Context) (tfsdk.Schema, diag.Diag
 					"**Important:** Even with an identical seed, it is not guaranteed that the same permutation " +
 					"will be produced across different versions of Terraform. This argument causes the " +
 					"result to be *less volatile*, but not fixed for all time.",
-				Type:          types.StringType,
-				Optional:      true,
-				PlanModifiers: []tfsdk.AttributePlanModifier{tfsdk.RequiresReplace()},
+				Type:     types.StringType,
+				Optional: true,
+				PlanModifiers: []tfsdk.AttributePlanModifier{
+					tfsdk.RequiresReplace(),
+				},
 			},
 			"input": {
 				Description: "The list of strings to shuffle.",
 				Type: types.ListType{
 					ElemType: types.StringType,
 				},
-				Required:      true,
-				PlanModifiers: []tfsdk.AttributePlanModifier{tfsdk.RequiresReplace()},
+				Required: true,
+				PlanModifiers: []tfsdk.AttributePlanModifier{
+					tfsdk.RequiresReplace(),
+				},
 			},
 			"result_count": {
 				Description: "The number of results to return. Defaults to the number of items in the " +
 					"`input` list. If fewer items are requested, some elements will be excluded from the " +
 					"result. If more items are requested, items will be repeated in the result but not more " +
 					"frequently than the number of items in the input list.",
-				Type:          types.Int64Type,
-				Optional:      true,
-				PlanModifiers: []tfsdk.AttributePlanModifier{tfsdk.RequiresReplace()},
+				Type:     types.Int64Type,
+				Optional: true,
+				PlanModifiers: []tfsdk.AttributePlanModifier{
+					tfsdk.RequiresReplace(),
+				},
 			},
 			"result": {
 				Description: "Random permutation of the list of strings given in `input`.",
@@ -69,18 +85,16 @@ func (r resourceShuffleType) GetSchema(context.Context) (tfsdk.Schema, diag.Diag
 	}, nil
 }
 
-func (r resourceShuffleType) NewResource(_ context.Context, p tfsdk.Provider) (tfsdk.Resource, diag.Diagnostics) {
-	return resourceShuffle{
-		p: *(p.(*provider)),
-	}, nil
+func (r *resourceType) NewResource(_ context.Context, p tfsdk.Provider) (tfsdk.Resource, diag.Diagnostics) {
+	return &resource{}, nil
 }
 
-type resourceShuffle struct {
-	p provider
-}
+var _ tfsdk.Resource = (*resource)(nil)
 
-func (r resourceShuffle) Create(ctx context.Context, req tfsdk.CreateResourceRequest, resp *tfsdk.CreateResourceResponse) {
-	var plan ShuffleModel
+type resource struct{}
+
+func (r *resource) Create(ctx context.Context, req tfsdk.CreateResourceRequest, resp *tfsdk.CreateResourceResponse) {
+	var plan modelV0
 	diags := req.Plan.Get(ctx, &plan)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
@@ -98,7 +112,7 @@ func (r resourceShuffle) Create(ctx context.Context, req tfsdk.CreateResourceReq
 	result := make([]attr.Value, 0, resultCount)
 
 	if len(input.Elems) > 0 {
-		rand := NewRand(seed)
+		rand := random.NewRand(seed)
 
 		// Keep producing permutations until we fill our result
 	Batches:
@@ -115,7 +129,7 @@ func (r resourceShuffle) Create(ctx context.Context, req tfsdk.CreateResourceReq
 		}
 	}
 
-	s := ShuffleModel{
+	s := modelV0{
 		ID:      types.String{Value: "-"},
 		Keepers: plan.Keepers,
 		Input:   plan.Input,
@@ -147,15 +161,24 @@ func (r resourceShuffle) Create(ctx context.Context, req tfsdk.CreateResourceReq
 }
 
 // Read does not need to perform any operations as the state in ReadResourceResponse is already populated.
-func (r resourceShuffle) Read(ctx context.Context, req tfsdk.ReadResourceRequest, resp *tfsdk.ReadResourceResponse) {
+func (r *resource) Read(ctx context.Context, req tfsdk.ReadResourceRequest, resp *tfsdk.ReadResourceResponse) {
 }
 
 // Update is intentionally left blank as all required and optional attributes force replacement of the resource
 // through the RequiresReplace AttributePlanModifier.
-func (r resourceShuffle) Update(ctx context.Context, req tfsdk.UpdateResourceRequest, resp *tfsdk.UpdateResourceResponse) {
+func (r *resource) Update(ctx context.Context, req tfsdk.UpdateResourceRequest, resp *tfsdk.UpdateResourceResponse) {
 }
 
 // Delete does not need to explicitly call resp.State.RemoveResource() as this is automatically handled by the
 // [framework](https://github.com/hashicorp/terraform-plugin-framework/pull/301).
-func (r resourceShuffle) Delete(ctx context.Context, req tfsdk.DeleteResourceRequest, resp *tfsdk.DeleteResourceResponse) {
+func (r *resource) Delete(ctx context.Context, req tfsdk.DeleteResourceRequest, resp *tfsdk.DeleteResourceResponse) {
+}
+
+type modelV0 struct {
+	ID          types.String `tfsdk:"id"`
+	Keepers     types.Map    `tfsdk:"keepers"`
+	Seed        types.String `tfsdk:"seed"`
+	Input       types.List   `tfsdk:"input"`
+	ResultCount types.Int64  `tfsdk:"result_count"`
+	Result      types.List   `tfsdk:"result"`
 }
